@@ -1,20 +1,19 @@
 buildscript {
     repositories {
+        google()
         mavenLocal()
         jcenter()
     }
     dependencies {
-        classpath "com.android.tools.build:gradle:2.3.3"
+        //noinspection GradleDependency
+        classpath "com.android.tools.build:gradle:3.2.1"
     }
 }
 
 repositories {
+    google()
     mavenLocal()
     jcenter()
-}
-
-task wrapper(type: Wrapper) {
-    gradleVersion = '4.0'
 }
 
 group = GROUP
@@ -27,12 +26,12 @@ sourceCompatibility = JavaVersion.VERSION_1_7
 targetCompatibility = JavaVersion.VERSION_1_7
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "26.0.0"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
 
     defaultConfig {
-        minSdkVersion 8
-        targetSdkVersion 25
+        minSdkVersion 16
+        targetSdkVersion 28
     }
 
     lintOptions {
@@ -45,7 +44,7 @@ android {
 }
 
 dependencies {
-    compile('com.android.support.test:runner:0.5')
+    implementation('androidx.test:rules:1.1.0')
 }
 
 afterEvaluate { project ->
