@@ -1,13 +1,3 @@
-/*
- * Copyright 2015 Shazam Entertainment Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
- *
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
- */
-
 package com.shazam.android.test.rules;
 
 import com.shazam.android.test.conditions.Condition;
@@ -25,6 +15,7 @@ import java.lang.reflect.Method;
  * if their {@link Condition Conditions} are satisfied.
  */
 public class ConditionalIgnoreRule implements TestRule {
+
     private static final Statement NO_OP_STATEMENT = new NoOpStatement();
 
     @Override
@@ -75,8 +66,6 @@ public class ConditionalIgnoreRule implements TestRule {
                 throw new RuntimeException("Ensure the no-arg constructor in your Condition class is public", e);
             }
         }
-
         return false;
     }
-
 }
