@@ -1,27 +1,20 @@
-/*
- * Copyright 2015 Shazam Entertainment Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
- *
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
- */
-
 package com.shazam.android.test.rules;
 
 import com.shazam.android.test.conditions.Condition;
 
-import java.lang.annotation.*;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
- * This annotation can be used on test cases or methods in order to exclude them in the test run on
- * the current device depending on its runtime.
- */
+ * This annotation can be used on test cases or methods in order to exclude
+ * them in the test run on the current device depending on its runtime.
+**/
 @Retention(value = RetentionPolicy.RUNTIME)
 @Target(value = {ElementType.METHOD, ElementType.TYPE})
 @Documented
 public @interface IgnoreWhen {
-
     Class<? extends Condition>[] device() default {};
 }
